requirements.txt
# opencv-python==3.4.3.18
opencv-python==4.1.1.26
tqdm==4.36.1
six==1.12.0
tensorflow-transform==0.15.0
apache-beam[gcp]==2.16.0
matplotlib==3.1.0
seaborn==0.9.0
scikit-learn==0.21.3
tf-explain==0.1.0